refactor: go back to a simpler settings structure
