+++ tests/settings.py
@@ -1,5 +1,7 @@
 """Made-up settings to use during tests."""
 
-GITHUB_BLENDED_PROJECT = "blendorg:42"
-GITHUB_OSPR_PROJECT = "testorg:17"
+# These should be in the in-memory form ready to patch into openedx_webhooks.settings
+
+GITHUB_BLENDED_PROJECT = ("blendorg", 42)
+GITHUB_OSPR_PROJECT = ("testorg", 17)
 GITHUB_PERSONAL_TOKEN = "github_pat_FooBarBaz"
