+++ tests/test_pull_request_opened.py
@@ -14,8 +14,8 @@ from openedx_webhooks.cla_check import (
     CLA_STATUS_NO_CONTRIBUTIONS,
     CLA_STATUS_PRIVATE,
 )
+from openedx_webhooks import settings
 from openedx_webhooks.gh_projects import pull_request_projects
-from openedx_webhooks.settings import settings
 from openedx_webhooks.tasks.github import pull_request_changed
 
 from .helpers import check_issue_link_in_markdown
